cleanup

+++ Common.CSharp/MathHelpers.cs
@@ -2,7 +2,7 @@
 {
     public static class MathHelpers
     {
-        public static uint GetGreatestCommonDivisor(uint a, uint b)
+        public static ulong GetGreatestCommonDivisor(ulong a, ulong b)
         {
             while (b != 0) (a, b) = (b, a % b);
             return a;
